+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 ## Unreleased
 
 - Accept `Closure|string` in `getParametersToResolve()`
-- Added `resolve(Closure)` to `ContainerInterface`
+- Added `resolve(callable)` to `ContainerInterface`
 
 ## 0.4.0
 ### 2023-04-27